Add vehicle section
